Fixed more bugs related to QPs over wands: in particular, missing background definitions (e.g. equality of summarising snapshots) and incorrect quantifiers (e.g. quantify over n-tuple instead of individual arguments)
